+++ cg/server/api.py
@@ -45,6 +45,7 @@ from cg.server.dto.orders.orders_request import OrdersRequest
 from cg.server.dto.orders.orders_response import Order, OrdersResponse
 from cg.server.dto.sequencing_metrics.sequencing_metrics_request import SequencingMetricsRequest
 from cg.server.ext import db, delivery_message_service, lims, order_service, osticket
+from cg.server.utils import parse_metrics_into_request
 from cg.store.models import (
     Analysis,
     Application,
@@ -53,7 +54,6 @@ from cg.store.models import (
     Customer,
     Pool,
     Sample,
-    SampleLaneSequencingMetrics,
     User,
     IlluminaSampleSequencingMetrics,
 )
@@ -374,23 +374,6 @@ def get_sequencing_metrics(flow_cell_name: str):
         )
     metrics_dtos: list[SequencingMetricsRequest] = parse_metrics_into_request(sequencing_metrics)
     return jsonify([metric.model_dump() for metric in metrics_dtos])
-
-
-def parse_metrics_into_request(
-    metrics: list[IlluminaSampleSequencingMetrics],
-) -> list[SequencingMetricsRequest]:
-    """Parse  sequencing metrics into a request."""
-    parsed_metrics: list[SequencingMetricsRequest] = []
-    for metric in metrics:
-        parsed_metric = SequencingMetricsRequest(
-            flow_cell_name=metric.instrument_run.device.internal_id,
-            flow_cell_lane_number=metric.flow_cell_lane,
-            sample_internal_id=metric.sample.internal_id,
-            sample_total_reads_in_lane=metric.total_reads_in_lane,
-            sample_base_percentage_passing_q30=metric.base_passing_q30_percent,
-        )
-        parsed_metrics.append(parsed_metric)
-    return parsed_metrics
 
 
 @BLUEPRINT.route("/analyses")
