+++ cg/server/dto/sequencing_metrics/sequencing_metrics_request.py
@@ -1,6 +1,6 @@
 """DTO for the sequencing metrics."""
 
-from pydantic import BaseModel, Field
+from pydantic import BaseModel
 
 
 class SequencingMetricsRequest(BaseModel):
