+++ tests/models/server/dto/test_sequencing_metrics.py
@@ -1,6 +1,6 @@
 """Tests for the sequencing metrics dtos."""
 
-from cg.server.api import parse_metrics_into_request
+from cg.server.utils import parse_metrics_into_request
 from cg.server.dto.sequencing_metrics.sequencing_metrics_request import SequencingMetricsRequest
 from cg.store.models import SampleLaneSequencingMetrics, IlluminaSampleSequencingMetrics
 from cg.store.store import Store
